+++ src/main/java/com/aca_disqo/generactive/repository/model/Stock.java
@@ -7,7 +7,7 @@ public class Stock extends Item {
     public Stock() {
     }
 
-    public Stock(int id, int basePrice, String name) {
+    public Stock(Long id, int basePrice, String name) {
         super(id, basePrice, name);
     }
